+++ src/components/AlignmentContainer/index.ts
@@ -1,0 +1,1 @@
+export * from './AlignmentContainer'
